+++ src/main/java/com/yahveh/dto/response/DetalleNotaEntregaResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -20,11 +21,11 @@ public class DetalleNotaEntregaResponse {
     private String lineaArticulo;
     private int codLinea;
     private int cantidad;
-    private float precioUnitario;
-    private float precioTotal;
-    private float precioSinFactura;
-    private float descuento;
-    private float precioConDescuento;
+    private BigDecimal precioUnitario;
+    private BigDecimal precioTotal;
+    private BigDecimal precioSinFactura;
+    private float descuento;          // porcentaje (0-100), no es dinero
+    private BigDecimal precioConDescuento;
     private int audUsuario;
     private LocalDateTime audFecha;
 }
