+++ src/main/java/com/yahveh/dto/response/NotaEntregaResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -28,6 +29,6 @@ public class NotaEntregaResponse {
     private String nombreEmpleado;
     private LocalDateTime audFecha;
     private List<DetalleNotaEntregaResponse> detalles;
-    private float totalGeneral;
+    private BigDecimal totalGeneral;
     private int totalArticulos;
 }
